Create a custom Docker image
The tools included in the Nextstrain base image (nextstrain/base) are
primarily meant for viral analyses. Bacterial analyses – in this case,
tuberculosis – require a different set of tools. While these tools may
eventually make their way into the Nextstrain base image, for now we
will maintain a separate image that extends nextstrain/base with the
necessary additions.

Conda environments are not the only way to add these tools – and they
aren't the most efficient use of Docker layers – but they are relatively
easy to maintain. The package recipes already handle the complex build
processes for these tools, which would otherwise be difficult to
reproduce in a Dockerfile.

See the new README file for details on building the image.